refactored code and fixed resetting score issue
