+++ .vscode/settings.json
@@ -14,5 +14,25 @@
     80,
     130,
     140,
-  ]
+  ],
+  "workbench.colorCustomizations": {
+    "activityBar.activeBackground": "#7295b1",
+    "activityBar.activeBorder": "#e8d6e0",
+    "activityBar.background": "#7295b1",
+    "activityBar.foreground": "#15202b",
+    "activityBar.inactiveForeground": "#15202b99",
+    "activityBarBadge.background": "#e8d6e0",
+    "activityBarBadge.foreground": "#15202b",
+    "statusBar.background": "#557c9b",
+    "statusBar.border": "#557c9b",
+    "statusBar.foreground": "#e7e7e7",
+    "statusBarItem.hoverBackground": "#7295b1",
+    "tab.activeBorder": "#7295b1",
+    "titleBar.activeBackground": "#557c9b",
+    "titleBar.activeForeground": "#e7e7e7",
+    "titleBar.border": "#557c9b",
+    "titleBar.inactiveBackground": "#557c9b99",
+    "titleBar.inactiveForeground": "#e7e7e799"
+  },
+  "peacock.color": "#557c9b"
 }
