+++ package.json
@@ -4,10 +4,12 @@
   "description": "An endless runner game based on Flappy Bird, made in Typescript with a local score system",
   "main": "script.js",
   "scripts": {
-    "dev": "lite-server",
-    "pretty": "prettier --config .prettierrc.json './*.ts' --write",
-    "lint": "eslint \"./*.ts\" --fix",
-    "quality": "yarn lint && yarn pretty"
+    "dev": "webpack-dev-server",
+    "pretty": "prettier --config .prettierrc.json './src/**/*.ts' --write",
+    "lint": "eslint \"./src/**/*.ts\" --fix",
+    "quality": "yarn lint && yarn pretty",
+    "build": "webpack",
+    "build:prod": "webpack --config webpack.config.prod.js"
   },
   "keywords": [
     "runner",
@@ -24,6 +26,8 @@
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^3.7.0",
     "@typescript-eslint/parser": "^3.7.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^4.0.0",
     "eslint": "^7.5.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-config-prettier": "^6.11.0",
@@ -31,7 +35,12 @@
     "eslint-plugin-prettier": "^3.1.4",
     "husky": "^4.2.5",
     "prettier": "2.0.5",
-    "typescript": "^3.9.7"
+    "style-loader": "^1.2.1",
+    "ts-loader": "^8.0.1",
+    "typescript": "^3.9.7",
+    "webpack": "^4.44.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   },
   "husky": {
     "hooks": {
