+++ src/scripts/constants.ts
@@ -1,0 +1,17 @@
+export const headerHeight = 80;
+
+export const __LEADERBOARD_KEY__ = "tsSRLB";
+
+export const __MUSIC_PREF_KEY__ = "tsSRMP";
+
+export const pipeColorPalette = [
+  "#f2002c", // red
+  "#fcf242", // yellow
+  "#1cc8f5", // blue
+  "#66ff00", // green
+  "#01f9c6", // teal
+  "#0ff0fc", // cyan
+  "#ffcf00", // orange
+  "#dfff11", // lime
+  "#bf00ff", // purple
+];
